Catch exceptions from backup runner
The backup runner context manager may throw BackupErrors.
The backupagent currently only checks for the success of the storage
save action and any exceptions, but not those from the runner
__enter__/__exit__ methods.

This issue was discovered when fixing backup unittests
in change: 312611
The unittests assert that the backup would be marked as FAILED, but
was marked as COMPLETED instead.

This patch just moves the 'with clause' of the backup runner to the
try-except block to intercept any exceptions originating from
the backup procedures.

Change-Id: I27d0cf1b9a174ca69eb82d26f14a4c0d03f75382
Partial-Bug: 1544522
(cherry picked from commit 3edcd5cdcf0d7d3928c67c1310a472df8d8af62a)