Delete replication snapshot when guestagent failed early.
In some extreme situations, like free memory is very limited, guestagent
may be very unstable. And if guestagent failed early in creating the
replication snapshot(before it's state change to BUILD), the replication
snapshot will stay in NEW and the instance will stay in BACKUP.

As when a backup(a replication snapshot is a backup too) is NEW, we can
safely delete it from the database.

Fixes: http://192.168.15.2/issues/10272

Change-Id: I71d1cef8467a2f03ac0c2a87fa69189186459da5
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>