Fix redis password with #
Currently, trove stripes in-line comments inside
trove.common.stream_codecs.PropertiesCodec, but it will cause a bug when
enabling root in redis with password like '@#$' or something including '#'
character.

First, add class property 'SUPPORT_INLINE_COMMENTS' in PropertiesCodec
which default value is True to control whether skipping in-line commnets
check or not. Then, We create a RedisPropertiesCodec from ProtiesCodec
but with property SUPPORT_INLINE_COMMENTS=False in order to skip the check
to support redis password with '#'.

Closes-bug: http://192.168.15.2/issues/11176
Signed-off-by: Fan Zhang <zh.f@outlook.com>