Fix creating MongoDB replica set from backup.
Before this patch, there were 3 main problems in creating MongoDB replica
set from backup:

1. create_cluster rpc call is not called, leaving the cluster and the
instances stay in 'BUILDING' status

The reason is in fact no overriding of TaskManager Manager has ever been
made. But fortunately there is a framework to do this kind of tricks:
a. add an non-existing rpc call (the name should begins with the datastore
   name, e.g. "mongo-");
b. add a mapping (as an entry to task_manager_manager_actions) for this
   rpc call to redirect TaskManager to the actual codes.

2. primary member restarts before prep_primary

This happens because of not correctly splitting primary and secondaries
in _create_replica_set.

Primary member shouldn't restart before prep_primary, because replica
set configuration has been saved during instance creating. When MongoDB
restarts, we'll get a uninitialized replica set, creating a admin user
will fail.

p.s. there are wrong block comments format in _create_replica_set. Block
comments should begin with "# ", "'''"(triple single quotes) is just a
multi-line string.

3. restoring primary member fails

The failure occurs at two places:
a. when restoring backup, mongorestore will parse admin authentication
info from the default credential file;
b. after being sucessfully restored, guestagent will connect to the
database to check whether a root user exists. Database connection needs
the admin authentication info from the default credential file.

But by the time of restoring, admin user and the credential file are not
created yet, so exceptions raise.

As when the admin user is not created and the database is not restarted
for the first time, all clients have administration permissions without
authentication, so when the credential file is not existing, the
authentication info could be safely set to empty.

Fixes: http://192.168.15.2/issues/11271

Change-Id: Ie2fe2833f87d070193298221a17fc4d1c2ca5b2c
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>